string lengths should be size_t, not int

+++ c_distance/hamming.c
@@ -19,8 +19,8 @@ static PyObject* distance(PyObject *self, PyObject *args)
         s2 = tmp;
     }
 
-    int s1_len = strlen(s1);
-    int s2_len = strlen(s2);
+    size_t s1_len = strlen(s1);
+    size_t s2_len = strlen(s2);
     long dist = s1_len - s2_len;
     for (size_t i = 0; i < s2_len; i++)
     {
